ENH: bring modules from kim_utils into KIMkit
Instead of having a seperate kim_utils package,
the modules needed for KIMkit will simply be included
in KIMkit.

Initally the thinking was that kim_utils would be a shared
dependency of the pipeline and KIMkit, but given the different
use cases, and small overlap of needed code, it makes more
sense to simply incorporate kimapi.py, kimcodes.py, and
kimobjects.py into KIMkit.

kimobjects.py contains the base classes of kim objects
that KIMkit objects inherit from. The KIMkit version
has commented out two properties of the  Runner class,
which will be unused in KIMkit, to remove a
dependency on our jinja templating code, which would expose
too many internal pipeline settings.

kimcodes.py is a superset of the features of the pipeline version,
with added functions to generate a new kimcode, and convert a
kimcode into a path in a KIMkit repository with the nested
directory structure based on subsections of the ID number.

kimapi.py is unchanged from the pipeline version, except for fixing
imports.