FEAT: initial feature commit to add mongodb support
Mongodb is now the database backend for KIMkit. Initially,
there will be 2 collections, "items" which stores the
metadata of KIMkit content, and "users", which stores information
about usernames for attribution purposes.

Major TODO is to add a sutible query interface, to allow
subsets of KIMkit content to be easily managed.